pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loadout"
version = "0.1.0"
description = "Workspace Context Switcher — record and replay developer workspace setups"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "PyYAML>=6.0",
]

[project.scripts]
loadout = "loadout.cli.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-timeout>=2.1",
]
firefox = [
    "lz4>=4.0",  # Required for reading Firefox session files
]
all = [
    "lz4>=4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["loadout*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 30
